# Optimise getUniqueByProp