@samwho 2023-12-14
@melohagan this function is eventually called through the `outputProcessing` function, which is already tested in a lot of different ways through other tests. On the backend side of Budibase we tend to test things through calls to the public API, rather than having granular unit tests.